% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_summary_diff_tool.R
\name{add_change_rate_cols}
\alias{add_change_rate_cols}
\title{add change rate cols}
\usage{
add_change_rate_cols(summary_df)
}
\arguments{
\item{summary_df}{}
}
\value{
data.frame
}
\description{
add cols which was mutated as change rate from past term data
(lag target col)  into the summary data
}
