% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_fillter.R
\name{filter_by_sd}
\alias{filter_by_sd}
\title{filter by sd}
\usage{
filter_by_sd(df, nse_col_name, n = 4, check = FALSE)
}
\arguments{
\item{df}{data.frame}

\item{nse_col_name}{対象となる列の列名}

\item{n}{標準偏差の何倍でフィルタするかを整数で指定。デフォルトは4}

\item{check}{Logical TRUEにすると、範囲外のデータを出力}
}
\value{
data.frame
}
\description{
対象となるデータ列の平均値からn倍標準偏差の位置で
フィルタリングする
}
