% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_summary_tool.R
\name{get_summary}
\alias{get_summary}
\title{get summary in a arbitrary numeric col}
\usage{
get_summary(df, nse_col)
}
\arguments{
\item{df}{data.frame}

\item{nse_col}{col-name(non-standard-eval)}
}
\value{
data.frame
}
\description{
get summary on numeric data of a arbitrary col.
}
\examples{
get_summary(data.frame(t = 1:10), t)
}
