% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_summary_by_group.R
\name{get_summary_by_group}
\alias{get_summary_by_group}
\title{get each summary a data by each group}
\usage{
get_summary_by_group(df, target_col_name, group_col_name)
}
\arguments{
\item{df}{data.frame}

\item{target_col_name}{target col name(non-standard-eval)}

\item{group_col_name}{which col as by group (non-standard-eval)}
}
\value{
data.frame
}
\description{
get each summary a target numeric data of arbitrary col
by each group.
}
\examples{
get_summary_by_group(data.frame(g = rep(c(1,2),10), v = 1:20), v, g)
}
