% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_summary_byTerm_tool.R
\name{get_summary_by_quarter}
\alias{get_summary_by_quarter}
\title{get summary by quarter}
\usage{
get_summary_by_quarter(df, nse_summary_col, nse_date_col)
}
\arguments{
\item{df}{data.frame}

\item{nse_summary_col}{target col which you like to get summary}

\item{nse_date_col}{date col}
}
\value{
data.frame
}
\description{
get summary of a col which you like by quarter.
}
