% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_summary_byTerm_tool.R
\name{get_summary_by_rolling_quarter}
\alias{get_summary_by_rolling_quarter}
\title{get rolling quarter summary}
\usage{
get_summary_by_rolling_quarter(df, nse_summary_col, nse_date_col)
}
\arguments{
\item{df}{data.frame}

\item{nse_summary_col}{terget col name which you like to know}

\item{nse_date_col}{terget date col name}
}
\value{
data.frame
}
\description{
get summary data by rolling quarter term.
}
