% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retiex_summary_tools.R
\name{retiex_rolling_summary}
\alias{retiex_rolling_summary}
\title{get rolling summary from reti data}
\usage{
retiex_rolling_summary(reti_data)
}
\arguments{
\item{reti_data}{}
}
\value{
data.frame
}
\description{
get rolling summary from reti data
}
