% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retiex_summary_tools.R
\name{retiex_summary}
\alias{retiex_summary}
\title{get summary from reti data}
\usage{
retiex_summary(reti_data, term = "q")
}
\arguments{
\item{reti_data}{data.frame}

\item{term}{string}
}
\value{
data.frame
}
\description{
get summary from reti data by year or quarter term.
}
