% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retiex_summary_by_time.R
\name{retiex_summary_by_time}
\alias{retiex_summary_by_time}
\title{get summary from reti data}
\usage{
retiex_summary_by_time(reti_data, time = "y")
}
\arguments{
\item{reti_data}{data.frame}

\item{time}{string}
}
\value{
data.frame
}
\description{
get summary from reti data by year or quarter term.
}
