% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_tool.R
\name{seq_date_by_qarter}
\alias{seq_date_by_qarter}
\title{make sequence date data by quarter}
\usage{
seq_date_by_qarter(min, max)
}
\arguments{
\item{min}{Date}

\item{max}{Date}
}
\value{
vector of Date
}
\description{
make sequence date data by quarter.
}
