% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_percent.R
\name{style_percent}
\alias{style_percent}
\title{Show human readable percent style}
\usage{
style_percent(num, keta = 1, diff = TRUE)
}
\arguments{
\item{num}{double}

\item{keta}{int}

\item{diff}{logical}
}
\value{
string
}
\description{
Change a double num into human readable percent style String.
}
\examples{
style_percent(0.1)
style_percent(0.1, keta = 2, diff = FALSE)
}
