% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_format.R
\name{style_quarter_date}
\alias{style_quarter_date}
\title{Show human readable quarter date style}
\usage{
style_quarter_date(date)
}
\arguments{
\item{date}{}
}
\value{
string
}
\description{
Change a Date data into
human readable quarter date style String.
}
\examples{
style_quarter_date(as.Date("2000-01-01"))
}
