% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_format.R
\name{style_rolling_term_quarter}
\alias{style_rolling_term_quarter}
\title{Show human readable rolling term style}
\usage{
style_rolling_term_quarter(date)
}
\arguments{
\item{date}{}
}
\value{
string
}
\description{
Change a Date data into human readable rolling term
by quarter date (1 year)style String.
}
\examples{
style_rolling_term_quarter(as.Date("2020-01-01"))
}
