% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_format.R
\name{style_yen}
\alias{style_yen}
\title{Show human readable yen style}
\usage{
style_yen(num, keta = 0, unit = "円")
}
\arguments{
\item{num}{int}

\item{keta}{int}

\item{unit}{string}
}
\value{
String
}
\description{
Change a num into human readable yen style String.
}
\examples{
style_yen(12345678)
style_yen(10000000, keta = 3, unit = "千円")
}
