% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_rolling_label.R
\name{transform_quarter_rolling_format}
\alias{transform_quarter_rolling_format}
\title{transform data for quarter rolling analysis}
\usage{
transform_quarter_rolling_format(df, nse_date_col)
}
\arguments{
\item{df}{data.frame}

\item{nse_date_col}{name of target date col}
}
\value{
data.frame
}
\description{
make pre data for quarter rolling analysis.
}
